chore: rename OnlineStatusNotification files to snake_case
