+++ src/nsn.ts
@@ -1,10 +1,10 @@
 export { useOnlineStatus } from './hooks'
 export type { NetworkInformation, OnlineStatusResult } from './hooks'
-export { default as OnlineStatusNotification } from './OnlineStatusNotification'
+export { default as OnlineStatusNotification } from './online_status_notification'
 export type {
   EventsCallback,
   OnlineStatusNotificationProps,
   OnlineStatusNotificationRef,
   Position,
   StatusText,
-} from './OnlineStatusNotification'
+} from './online_status_notification'
